Move #6: 모든 컴포넌트들 component 패키지로 이동
